Removed PHPCrawl from core repo, pulling it in with Composer (Fixes #6)
The PHPCrawl library is not maintained by us, and it also has a different license, so pulling it in via Composer
makes much more sense than copying the files in here.  Unfortunately, the repository hasn't been put into Packagist
by the maintainer yet, so in the meantime I have requested this with them, and set it up myself.