FIX: Keep inferred URLs separate do reduce duplicates.
Before this change, there was a bug, that if a parent URL didn't exist prior to processing, but did exist after it, a duplicate would be created.

This change keeps so-called "inferred" URLs separate from directly spidered ones, and recalculates the necessary inferred URLs after a processing function is applied.