NEW Possibility to trim content from selectors
This is done by adding an "Exclude CSS Selector" field to the ImportRules. This exclude selector trims aways those selectors found inside the css selector.

An example is the site you are scraping are having a div#content where all content is. Inside this content there are also an h1 with the page title. To be able to exlude the h1 from the Content field you would "h1" to the exclude css selectors.